all encoders/decoders working properly, need copy functionality.
