+++ Pinpoint.Plugin.EncodeDecode/EncodeDecodeResult.cs
@@ -1,0 +1,19 @@
+﻿using FontAwesome5;
+using Pinpoint.Core;
+
+namespace Pinpoint.Plugin.EncodeDecode
+{
+    public class EncodeDecodeResult: AbstractFontAwesomeQueryResult
+    {
+        public EncodeDecodeResult(string result): base(result)
+        {
+            
+        }
+        public override void OnSelect()
+        {
+            
+        }
+
+        public override EFontAwesomeIcon FontAwesomeIcon => EFontAwesomeIcon.Solid_Hashtag;
+    }
+}
