+++ Pinpoint.Plugin.EncodeDecode/IEncodeDecodeHandler.cs
@@ -1,0 +1,8 @@
+﻿namespace Pinpoint.Plugin.EncodeDecode
+{
+    public interface IEncodeDecodeHandler
+    {
+        string Encode(string stringToEncode);
+        string Decode(string stringToDecode);
+    }
+}
